+++ activerecord/lib/active_record/connection_adapters/sqlite3/quoting.rb
@@ -20,6 +20,10 @@ module ActiveRecord
           @quoted_column_names[name] ||= %Q("#{super.gsub('"', '""')}")
         end
 
+        def quote_array(a)
+          raise TypeError, "can't quote Array"
+        end
+
         def quoted_time(value)
           value = value.change(year: 2000, month: 1, day: 1)
           quoted_date(value).sub(/\A\d\d\d\d-\d\d-\d\d /, "2000-01-01 ")